+++ tests/algorithms/extended_temporal_memory/etm_sequence_memory_test.py
@@ -1,6 +1,6 @@
 # ----------------------------------------------------------------------
 # Numenta Platform for Intelligent Computing (NuPIC)
-# Copyright (C) 2017, Numenta, Inc.  Unless you have an agreement
+# Copyright (C) 2016-2017, Numenta, Inc.  Unless you have an agreement
 # with Numenta, Inc., for a separate license for this software code, the
 # following terms and conditions apply:
 #
@@ -20,32 +20,30 @@
 # ----------------------------------------------------------------------
 
 """
-Run the apical tiebreak tests on the C++ ExtendedTemporalMemory.
+Run the sequence memory tests on the C++ ExtendedTemporalMemory.
 """
 
 import unittest
 
 from nupic.bindings.experimental import ExtendedTemporalMemory
-from htmresearch.support.shared_tests.apical_tiebreak_test_base import (
-  ApicalTiebreakTestBase)
+
+from htmresearch.support.shared_tests.sequence_memory_test_base import (
+  SequenceMemoryTestBase)
 
 
-class ExtendedTMCPP_ApicalTiebreakTests(ApicalTiebreakTestBase,
-                                        unittest.TestCase):
+class ExtendedTM_SequenceMemoryTests(SequenceMemoryTestBase,
+                                     unittest.TestCase):
   """
-  Run the apical tiebreak tests on the C++ ExtendedTemporalMemory.
+  Run the sequence memory tests on the C++ ExtendedTemporalMemory.
   """
 
-  def constructTM(self, columnCount, basalInputSize, apicalInputSize,
-                  cellsPerColumn, initialPermanence, connectedPermanence,
-                  minThreshold, sampleSize, permanenceIncrement,
-                  permanenceDecrement, predictedSegmentDecrement,
-                  activationThreshold, seed):
+  def constructTM(self, columnCount, cellsPerColumn, initialPermanence,
+                  connectedPermanence, minThreshold, sampleSize,
+                  permanenceIncrement, permanenceDecrement,
+                  predictedSegmentDecrement, activationThreshold, seed):
 
     params = {
       "columnDimensions": (columnCount,),
-      "basalInputDimensions": (basalInputSize,),
-      "apicalInputDimensions": (apicalInputSize,),
       "cellsPerColumn": cellsPerColumn,
       "initialPermanence": initialPermanence,
       "connectedPermanence": connectedPermanence,
@@ -57,29 +55,20 @@ class ExtendedTMCPP_ApicalTiebreakTests(ApicalTiebreakTestBase,
       "activationThreshold": activationThreshold,
       "seed": seed,
       "learnOnOneCell": False,
-      "formInternalBasalConnections": False,
     }
 
     self.tm = ExtendedTemporalMemory(**params)
 
 
-  def compute(self, activeColumns, basalInput, apicalInput, learn):
-
-    activeColumns = sorted(activeColumns)
-    basalInput = sorted(basalInput)
-    apicalInput = sorted(apicalInput)
-
+  def compute(self, activeColumns, learn):
     # Use depolarizeCells + activateCells rather than tm.compute so that
     # getPredictiveCells returns predictions for the current timestep.
-    self.tm.depolarizeCells(activeCellsExternalBasal=basalInput,
-                            activeCellsExternalApical=apicalInput,
-                            learn=learn)
-    self.tm.activateCells(activeColumns,
-                          reinforceCandidatesExternalBasal=basalInput,
-                          growthCandidatesExternalBasal=basalInput,
-                          reinforceCandidatesExternalApical=apicalInput,
-                          growthCandidatesExternalApical=apicalInput,
-                          learn=learn)
+    self.tm.depolarizeCells(learn=learn)
+    self.tm.activateCells(sorted(activeColumns), learn=learn)
+
+
+  def reset(self):
+    self.tm.reset()
 
 
   def getActiveCells(self):
