--- tests/algorithms/extended_temporal_memory/etm_unit_cpp_test.py
@@ -1,35 +1,0 @@
-#!/usr/bin/env python
-# ----------------------------------------------------------------------
-# Numenta Platform for Intelligent Computing (NuPIC)
-# Copyright (C) 2016, Numenta, Inc.  Unless you have an agreement
-# with Numenta, Inc., for a separate license for this software code, the
-# following terms and conditions apply:
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero Public License version 3 as
-# published by the Free Software Foundation.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-# See the GNU Affero Public License for more details.
-#
-# You should have received a copy of the GNU Affero Public License
-# along with this program.  If not, see http://www.gnu.org/licenses.
-#
-# http://numenta.org/licenses/
-# ----------------------------------------------------------------------
-
-import unittest
-
-from etm_unit_test_base import ExtendedTemporalMemoryUnitTest
-import nupic.bindings.experimental
-
-
-
-class ExtendedTemporalMemoryUnitTestCPP(ExtendedTemporalMemoryUnitTest,
-                                        unittest.TestCase):
-
-
-  def getTMClass(self):
-    return nupic.bindings.experimental.ExtendedTemporalMemory
